break vesting pool monolith into smaller tests
